+++ test/synopsis.rb
@@ -6,13 +6,13 @@ LINE = '-rw-r--r-- 1 user users   87 Jun 18 18:16 CHANGELOG.md'
 WANT = ['-rw-r--r--', '1', 'user', 'users', '87', 'Jun 18 18:16', 'CHANGELOG.md']
 
 describe 'synopsis' do
-  it 'allows `remove` and `limit` to be combined' do
+  it 'allows `remove_empty` and `limit` to be combined' do
     s = StringSplitter.new(remove_empty: true)
 
-    result = s.split(':foo:bar:baz:quux:', ':', at: 1)
+    result = s.split('::foo:bar::baz:quux::', ':', at: 1)
     assert { result == ['foo', 'bar:baz:quux'] }
 
-    result = s.split(':foo:bar:baz:quux:', ':') { |i| i == 1 }
+    result = s.split('::foo:bar::baz:quux::', ':') { |split| split.index == 1 }
     assert { result == ['foo', 'bar:baz:quux'] }
   end