+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0 - TBD
+
+- **breaking change**: the block now takes a single `split` object with an `index`
+  field, rather than seperate `index` and `split` arguments
+- add support for negative indices in the value supplied to the `at` option
+- add a `count` field to the split object containing the total number of splits
+
 ## 0.0.1 - 2018-06-21
 
-* initial release
+- initial release